services/api/pyproject.toml
[project]
name = "api"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [{ name = "Leon Weimann", email = "nachhilfe@leonweimann.de" }]
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.116.1",
    "gaebio",
    "passlib[bcrypt]>=1.7.4",
    "argon2-cffi>=21.3.0",
    "python-jose[cryptography]>=3.5.0",
    "python-multipart>=0.0.20",
    "uvicorn>=0.35.0",
    "sqlalchemy>=2.0.44",
    "databases[sqlite]>=0.9.0",
    "alembic>=1.17.1",
]

[build-system]
requires = ["uv_build>=0.8.17,<0.9.0"]
build-backend = "uv_build"

[tool.uv.sources]
gaebio = { workspace = true }
